Modularize the main.go file and process the domains.txt file for the ./subdomain-checker domains.txt command to support domains starting with http(s)://.
